+++ app/src/main/java/eu/kanade/domain/source/service/SourcePreferences.kt
@@ -94,6 +94,8 @@ class SourcePreferences(
 
     // fun sourcesTabSourcesInCategories() = preferenceStore.getStringSet("sources_tab_source_categories", mutableSetOf())
 
+    fun showBrokenMangaSources() = preferenceStore.getBoolean("show_broken_manga_sources", false)
+
     fun dataSaver() = preferenceStore.getEnum("data_saver", DataSaver.NONE)
 
     fun dataSaverIgnoreJpeg() = preferenceStore.getBoolean("ignore_jpeg", false)